Disable running queries while paused
This is necessary because SQLite currently holds a lock on the full
connection while running tracing callbacks, which causes a deadlock when
attempting to execute a query on a paused connection.

Fixing this would require further modification to SQLite; otherwise a
(tedious) workaround is to pause the app with the C# debugger, freeze the
querying thread(s) outside of executing any queries, and let the debugger
thread run.